veyron/lib/exec: Fix panic in handle.Clean()
Fixes a panic in handle.Clean() when Start() fails.

panic: runtime error: invalid memory address or nil pointer dereference
[...]
Feb 06 11:19:07 rpi001 agent_deviced.sh[419]: v.io/core/veyron/lib/exec.(*ParentHandle).Kill(0x15dfdd80, 0x0, 0x0)
Feb 06 11:19:07 rpi001 agent_deviced.sh[419]: /home/rthellend/vanadium/release/go/src/v.io/core/veyron/lib/exec/parent.go:324 +0x38
Feb 06 11:19:07 rpi001 agent_deviced.sh[419]: v.io/core/veyron/lib/exec.(*ParentHandle).Clean(0x15dfdd80, 0x0, 0x0)
Feb 06 11:19:07 rpi001 agent_deviced.sh[419]: /home/rthellend/vanadium/release/go/src/v.io/core/veyron/lib/exec/parent.go:334 +0x30

+++ lib/exec/parent.go
@@ -321,11 +321,17 @@ func (p *ParentHandle) Exists() bool {
 
 // Kill kills the child process.
 func (p *ParentHandle) Kill() error {
+	if p.c.Process == nil {
+		return errors.New("no such process")
+	}
 	return p.c.Process.Kill()
 }
 
 // Signal sends the given signal to the child process.
 func (p *ParentHandle) Signal(sig syscall.Signal) error {
+	if p.c.Process == nil {
+		return errors.New("no such process")
+	}
 	return syscall.Kill(p.c.Process.Pid, sig)
 }
